requirements.txt
holidays==0.63
pydrive==1.3.1
airflow-provider-kafka==0.2.1
xlsxwriter==3.0.6
# confluent-kafka==1.8.2
python-gnupg==0.4.5
pendulum
openpyxl==3.1.2
ruamel.yaml==0.17.21
prettytable==3.6.0
weasyprint==58.1
redash-api-client==0.3.0
pydrive2==1.15.4
xlrd==2.0.1
pdfkit==1.0.0
pikepdf==8.6.0
# logger==1.4
# matplotlib==3.7.1
astronomer-starship==2.0.5
apache-airflow-providers-amazon==9.1.0
apache-airflow-providers-snowflake
polars>=0.20.16,<0.21
s3fs>=2023.6.0